Changed how runtime errors such as division by zero are reported.
We used to allow operations to be triggered by resources and we used
this for reporting common errors through the builtin exception instance
err. This trick does not work anymore, so I changed things around.

First I added Error.exc so that err#raise can report an exception.
Then in pervasives I added a type of runtime errors, changed err so
that it accepts runtime errors (rather than strings), and wrote
convenience handlers for handling division by zero and others.

So now division by zero is handled as

  with divisionByZero x with ...

which will evaluate to x if err#raise DivisionByZero happens.