Updated tests to match the new behavior of non-interactive files.
There is now a difference between loading a file with -l and loading
it as an anonymous command-line argument. With -l the file is loaded
silently (interactive mode is off), while in the other mode we print
out toplevel values and definitions, just like we would at toplevel.

Consequently, we can get rid of top-level "checks" and "check_vals"
in tests, but more importantly also in examples.