Better autoconf setup, got rid of the shell wrapper.
The shell wrapper which looked for rlwrap and ledit is gone, now
eff does this by itself, but only if the interactive shell is
started. Should eff report that a line editor was found?

Makefile.am now declares the eff executable as the main target
(rather than the shell wrapper), as it should. The trick with
settting eff_SOURCES to the empty value is needed, orelse automake
thinks eff is a C program and wants a C compiler.