Fixed descriptions of instances.
Each instance now carries an optional description that we use when showing the
instance. Built-in instances are described by "standard I/O", "standard error",
and "random number generation". Instances created with new have no descriptions.
If an instance has a description, we show them as <description>, if not, we show
them as <instance #...>.

Beforehand, each instance already carried a description, but this was never
used, and each instance showed as <instance #...>, even the built-in ones.