Code review of pattern exhaustive checking. There are problems.
There are at least two problems:

1. The code assumed all types are inhabited, but this is not the case for the empty type.

2. Exhaustiveness checking is called inside infer.ml, so it can happen quite easily that
pattern warning are issued before type errors. The exhaustiveness code is quite explicit
that it should be called *after* type checking because it assume everything is well typed.

I am inclined to suggest that the inference phase should emit fully typed code, this will be useful in general,
for example when we want to relax the value restriction.

And exhaustiveness checking really has to traverse the code separately from inference, there is no way around that.