Changed optional positions in error messages to mandatory. Added checking of shadowing of labels, fields, and operations.
We really should strive to always equip error messages with positions. This seems to be doable, except for
runtime errors. In fact, in the long run it is going to be almost impossible to always equip runtime errors
with source positions (think of compiled code). Perhaps a good strategy would be to require positions on
syntax and type errors, and to remove positions from runtime errors (and exceptions).

As for shadowing, please note that eff MUST check for shadowing of labels, fields and operations because
it uses them during runtime. Shadowing of any of these breaks type safety.