Put Big_int back in.
The switch to int from Big_int did not actually solve any problems. It just made
it look like we could use ocaml = and < where we actually could not, because we
still needed custom comparison functions because of comparison of instances.
So, let us not delete existing functionality in the name of simplification.