Separated types and type definitions.
As it turns out, there are no terms with types Sum, Record, or Effect.
Instead, those terms are of the type Apply (name, ...), where name is
the name of the defined type. The only proper way is of course to separate
types and type definitions.

Type definitions are Sum, Record, Effect and Inline. The first three are
more or less what they were before, while the last one represents type
abbreviations. This separation introduces a few additional functions
(we now need subst_tydef and fresh_tydef), but removes a lot of unused cases in
Type.free_params, Print.type, and Unify.solve. Some of them were already
"assert false"s and others them provided a computation that was never called.