Added constraints. Does compile, but does not work.
Each type scheme is equipped with a list of constraints (inequalities between types,  dirts, or regions).
Unify.solve now returns the remaining constraints in addition to the substitution it managed to perform.
Constraints can be substituted, refreshed, and similar, but there is a bug when you perform a substitution on a context.